add plan preview to plan detail screen
