+++ app/app/devotional_detail/[planId]/index.tsx
@@ -1,14 +1,20 @@
 import { useFetchDevotionalPlanById } from '@/src/hooks/useDevotionalPlans';
-import { useMyPlanProgressPlans } from '@/src/hooks/usePlanProgress';
+import {
+  useDayItemTemplates,
+  useDevotionalDays,
+  useMyPlanProgressPlans,
+  useStartPlanProgress,
+} from '@/src/hooks/usePlanProgress';
 import { useSavedPlans, useToggleSavedPlan } from '@/src/hooks/useSavedPlans';
 import DevotionalDetailScreen from '@/src/screens/DevotionalDetailScreen';
 import { useAuth } from '@/src/state/AuthContext';
+import { sortDayItems } from '@/src/utils';
 import BottomSheet from '@gorhom/bottom-sheet';
-import { useMemo, useRef } from 'react';
+import { useEffect, useMemo, useRef, useState } from 'react';
 
 import { useTogglePlanReaction } from '@/src/hooks/usePlanReactions';
 import { Stack, useLocalSearchParams, useRouter } from 'expo-router';
-import { Platform, useColorScheme } from 'react-native';
+import { Alert, Platform, useColorScheme } from 'react-native';
 
 export default function DevotionalDetail() {
   const { planId } = useLocalSearchParams<{ planId: string }>();
@@ -28,9 +34,19 @@ export default function DevotionalDetail() {
 
   const router = useRouter();
   const colorScheme = useColorScheme();
+  const startPlanProgressMutation = useStartPlanProgress();
   const myPlanProgressPlansQuery = useMyPlanProgressPlans(session?.user.id);
   const planQuery = useFetchDevotionalPlanById(planId);
   const plan = planQuery.data;
+  const daysQuery = useDevotionalDays(planId);
+  const previewDays = useMemo(() => daysQuery.data ?? [], [daysQuery.data]);
+  const [selectedPreviewDay, setSelectedPreviewDay] = useState(1);
+  const selectedPreviewDayData = previewDays.find((day) => day.day_number === selectedPreviewDay);
+  const dayItemsQuery = useDayItemTemplates(planId, selectedPreviewDayData?.id ?? '');
+  const previewItems = useMemo(
+    () => [...(dayItemsQuery.data ?? [])].sort(sortDayItems),
+    [dayItemsQuery.data],
+  );
   const isPrivatePlan = plan?.visibility === 'private';
   const canStartPlan = !isPrivatePlan || plan?.author_id === session?.user.id;
   const activePlanProgresses = useMemo(
@@ -52,6 +68,16 @@ export default function DevotionalDetail() {
     helpful_count: plan?.helpful_count ?? 0,
     user_reaction: plan?.user_reaction === 'helpful' ? ('helpful' as const) : null,
   };
+
+  useEffect(() => {
+    if (!previewDays.length) return;
+
+    const selectedDayExists = previewDays.some((day) => day.day_number === selectedPreviewDay);
+    if (!selectedDayExists) {
+      setSelectedPreviewDay(previewDays[0].day_number);
+    }
+  }, [previewDays, selectedPreviewDay]);
+
   const handleToggleReaction = () => {
     if (isGuest) {
       router.push('/app/(auth)/signin');
@@ -66,6 +92,44 @@ export default function DevotionalDetail() {
     }
     reportSheetRef.current?.expand();
   };
+  const handleContinuePress = () => {
+    if (!currentActivePlanProgress?.progress_id) return;
+
+    router.push(`/app/plan_progress/${currentActivePlanProgress.progress_id}`);
+  };
+
+  const handleStartPress = (mode: 'solo' | 'group') => {
+    if (startPlanProgressMutation.isPending) return;
+
+    if (isGuest) {
+      router.push('/(auth)/signin');
+      return;
+    }
+
+    if (plan?.visibility === 'private' && plan.author_id !== session?.user.id) {
+      handleContinuePress();
+      return;
+    }
+
+    if (mode === 'group') {
+      router.push(`/app/devotional_detail/${planId}/start-date`);
+      return;
+    }
+
+    startPlanProgressMutation.mutate(
+      { plan_id: planId, user_id: session?.user.id! },
+      {
+        onSuccess: (progress) => router.push(`/app/plan_progress/${progress.id}`),
+        onError: (error) => {
+          Alert.alert(
+            'Could not start plan',
+            error instanceof Error ? error.message : 'Please try again.',
+          );
+        },
+      },
+    );
+  };
+
   return (
     <>
       <Stack.Screen
@@ -93,15 +157,19 @@ export default function DevotionalDetail() {
         currentReaction={currentReaction}
         plan={planQuery.data}
         reportSheetRef={reportSheetRef}
-        isLoading={planQuery.isLoading || myPlanProgressPlansQuery.isLoading}
+        isLoading={planQuery.isLoading || myPlanProgressPlansQuery.isLoading || daysQuery.isLoading}
+        previewDays={previewDays}
+        selectedPreviewDay={selectedPreviewDay}
+        selectedPreviewDayData={selectedPreviewDayData}
+        previewItems={previewItems}
+        previewItemsLoading={dayItemsQuery.isLoading}
+        hasActiveSoloPlanProgress={!!currentSoloPlanProgress?.progress_id}
         hasActivePlanProgress={!!currentActivePlanProgress?.progress_id}
         canStartPlan={!!canStartPlan}
         isPrivatePlan={!!isPrivatePlan}
-        onContinuePress={() => {
-          if (!currentActivePlanProgress?.progress_id) return;
-
-          router.push(`/app/plan_progress/${currentActivePlanProgress.progress_id}`);
-        }}
+        isStartingSoloPlan={startPlanProgressMutation.isPending}
+        onSelectPreviewDay={setSelectedPreviewDay}
+        onContinuePress={handleContinuePress}
         isSaved={isSaved}
         onToggleSave={() => {
           if (isGuest) {
@@ -110,39 +178,7 @@ export default function DevotionalDetail() {
           }
           toggleSavedPlan(planId, isSaved, plan ?? undefined);
         }}
-        onStartPress={(mode: 'solo' | 'group') => {
-          if (startPlanProgressMutation.isPending) return;
-
-          if (isGuest) {
-            router.push('/app/(auth)/signin');
-            return;
-          }
-
-          if (plan?.visibility === 'private' && plan.author_id !== session?.user.id) {
-            if (currentActivePlanProgress?.progress_id) {
-              router.push(`/app/plan_progress/${currentActivePlanProgress.progress_id}`);
-            }
-            return;
-          }
-
-          if (mode === 'group') {
-            router.push(`/app/devotional_detail/${planId}/start-date`);
-            return;
-          }
-
-          startPlanProgressMutation.mutate(
-            { plan_id: planId, user_id: session?.user.id! },
-            {
-              onSuccess: (progress) => router.push(`/app/plan_progress/${progress.id}`),
-              onError: (error) => {
-                Alert.alert(
-                  'Could not start plan',
-                  error instanceof Error ? error.message : 'Please try again.',
-                );
-              },
-            },
-          );
-        }}
+        onStartPress={handleStartPress}
       />
     </>
   );
